#113 test: マスタ系のフォームのテストを抽象化する
